Add files via upload
The raindrops are supposed to be a grid not a row, the grid wont show. The code works when I set the the raindrops to move right off the screen, but fails when I tell it to move down and I have no idea why. I changed the x to y in the raindrop class and it wont work now.